post reaction 보여주는 cell도 'postReactions' feature로 분리
