+++ src/widgets/post/ui/PostWidget.tsx
@@ -17,7 +17,7 @@ import {
   TableRow,
   Textarea,
 } from "@/shared/ui"
-import { Edit2, MessageSquare, Plus, Search, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react"
+import { Edit2, MessageSquare, Plus, Search, ThumbsUp, Trash2 } from "lucide-react"
 import { useEffect, useState } from "react"
 import { useLocation, useNavigate } from "react-router-dom"
 
@@ -30,6 +30,8 @@ import { SortBy } from "@/features/sortBy/ui/SortBy.tsx"
 import { SortOrder } from "@/features/sortOrder/ui/SortOrder.tsx"
 import { Pagination } from "@/features/pagination/ui/Pagination.tsx"
 import { PostTags } from "@/features/postTags/ui/PostTags.tsx"
+import { PostAuthor } from "@/features/postAuthor/ui/PostTags.tsx"
+import { PostReactions } from "@/features/PostReactions/ui/PostReactions.tsx"
 
 export interface NewComment {
   body: string
@@ -131,11 +133,6 @@ const likeCommentAPI = async (id: number, likes: number): Promise<Comment> => {
   return response.json()
 }
 
-const fetchUserAPI = async (userId: number): Promise<UserProfile> => {
-  const response = await fetch(`/api/users/${userId}`)
-  return response.json()
-}
-
 export const PostWidget = () => {
   const navigate = useNavigate()
   const location = useLocation()
@@ -157,8 +154,6 @@ export const PostWidget = () => {
   const [comments, setComments] = useState<Record<number, Comment[]>>({})
   const [selectedComment, setSelectedComment] = useState<Comment | null>(null)
   const [newComment, setNewComment] = useState<NewComment>(defaultNewComment)
-
-  const [selectedUser, setSelectedUser] = useState<UserProfile | null>(null)
 
   const [showEditDialog, setShowEditDialog] = useState<boolean>(false)
   const [showAddCommentDialog, setShowAddCommentDialog] = useState<boolean>(false)
@@ -350,17 +345,6 @@ export const PostWidget = () => {
     setShowPostDetailDialog(true)
   }
 
-  // 사용자 모달 열기
-  const openUserModal = async (user: UserProfile | undefined) => {
-    if (!user) return
-    try {
-      const userData = await fetchUserAPI(user.id)
-      setSelectedUser(userData)
-    } catch (error) {
-      console.error("사용자 정보 가져오기 오류:", error)
-    }
-  }
-
   useEffect(() => {
     fetchPosts()
     updateURL()
@@ -488,21 +472,10 @@ export const PostWidget = () => {
                       </div>
                     </TableCell>
                     <TableCell>
-                      <div
-                        className="flex items-center space-x-2 cursor-pointer"
-                        onClick={() => openUserModal(post.author)}
-                      >
-                        <img src={post.author?.image} alt={post.author?.username} className="w-8 h-8 rounded-full" />
-                        <span>{post.author?.username}</span>
-                      </div>
+                      <PostAuthor post={post} />
                     </TableCell>
                     <TableCell>
-                      <div className="flex items-center gap-2">
-                        <ThumbsUp className="w-4 h-4" />
-                        <span>{post.reactions?.likes || 0}</span>
-                        <ThumbsDown className="w-4 h-4" />
-                        <span>{post.reactions?.dislikes || 0}</span>
-                      </div>
+                      <PostReactions reactions={post.reactions} />
                     </TableCell>
                     <TableCell>
                       <div className="flex items-center gap-2">
@@ -610,40 +583,6 @@ export const PostWidget = () => {
           <div className="space-y-4">
             <p>{highlightText(selectedPost?.body || "", searchQuery)}</p>
             {renderComments(selectedPost?.id || 0)}
-          </div>
-        </DialogContent>
-      </Dialog>
-
-      {/* 사용자 모달 */}
-      <Dialog open={!!selectedUser} onOpenChange={(open) => !open && setSelectedUser(null)}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>사용자 정보</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <img src={selectedUser?.image} alt={selectedUser?.username} className="w-24 h-24 rounded-full mx-auto" />
-            <h3 className="text-xl font-semibold text-center">{selectedUser?.username}</h3>
-            <div className="space-y-2">
-              <p>
-                <strong>이름:</strong> {selectedUser?.firstName} {selectedUser?.lastName}
-              </p>
-              <p>
-                <strong>나이:</strong> {selectedUser?.age}
-              </p>
-              <p>
-                <strong>이메일:</strong> {selectedUser?.email}
-              </p>
-              <p>
-                <strong>전화번호:</strong> {selectedUser?.phone}
-              </p>
-              <p>
-                <strong>주소:</strong> {selectedUser?.address?.address}, {selectedUser?.address?.city},{" "}
-                {selectedUser?.address?.state}
-              </p>
-              <p>
-                <strong>직장:</strong> {selectedUser?.company?.name} - {selectedUser?.company?.title}
-              </p>
-            </div>
           </div>
         </DialogContent>
       </Dialog>
